すでにBrewfilwがある場合の処理を作成

+++ .config/.zshrc
@@ -184,11 +184,15 @@ if [ -f "$HOME/.env" ]; then
     if [ "$LOCAL_NAME" = "macbook" ]; then
 		bfile='/Users/atsuki/ghq/github.com/maro114510/dotfiles/mac_book/Brewfile' 
 		alias brewd="brew bundle dump --force --file=$bifle"
-		ln -sfn "$bfile" "$HOME/Brewfile"
+		if [ -f "$HOME/Brewfile" ]; then
+			ln -sfn "$bfile" "$HOME/Brewfile"
+		fi
     elif [ "$LOCAL_NAME" = "macmini" ]; then
 		bfile='/Users/atsuki/ghq/github.com/maro114510/dotfiles/mac_mini/Brewfile' 
 		alias brewd="brew bundle dump --force --file=$bfile"
-		ln -sfn "$bfile" "$HOME/Brewfile"
+		if [ -f "$HOME/Brewfile" ]; then
+			ln -sfn "$bfile" "$HOME/Brewfile"
+		fi
     fi
 fi
